+++ src/v0/destinations/snapchat_conversion/util.js
@@ -1,14 +1,17 @@
 const get = require('get-value');
 const sha256 = require('sha256');
+const { InstrumentationError, ConfigurationError } = require('@rudderstack/integrations-lib');
+const moment = require('moment/moment');
 const logger = require('../../../logger');
 
 const {
   isDefinedAndNotNull,
   getFieldValueFromMessage,
   defaultBatchRequestConfig,
+  getValidDynamicFormConfig,
 } = require('../../util');
-const { ENDPOINT } = require('./config');
 const { JSON_MIME_TYPE } = require('../../util/constant');
+const { ENDPOINT } = require('./config');
 
 const channelMapping = {
   web: 'WEB',
@@ -118,7 +121,7 @@ function generateBatchedPayloadForArray(events, destination) {
     batch: JSON.stringify(batchResponseList),
   };
 
-  batchedRequest.endpoint = ENDPOINT;
+  batchedRequest.endpoint = ENDPOINT.Endpoint_v2;
   batchedRequest.headers = {
     'Content-Type': JSON_MIME_TYPE,
     Authorization: `Bearer ${apiKey}`,
@@ -127,6 +130,83 @@ function generateBatchedPayloadForArray(events, destination) {
   return batchedRequest;
 }
 
+// Checks if there are any mapping events for the track event and returns them
+const eventMappingHandler = (message, destination) => {
+  let event = get(message, 'event');
+
+  if (!event) {
+    throw new InstrumentationError('Event name is required');
+  }
+  event = event.toString().trim().replace(/\s+/g, '_');
+
+  let { rudderEventsToSnapEvents } = destination.Config;
+  const mappedEvents = new Set();
+
+  if (Array.isArray(rudderEventsToSnapEvents)) {
+    rudderEventsToSnapEvents = getValidDynamicFormConfig(
+      rudderEventsToSnapEvents,
+      'from',
+      'to',
+      'snapchat_conversion',
+      destination.ID,
+    );
+    rudderEventsToSnapEvents.forEach((mapping) => {
+      if (mapping.from.trim().replace(/\s+/g, '_').toLowerCase() === event.toLowerCase()) {
+        mappedEvents.add(mapping.to);
+      }
+    });
+  }
+
+  return [...mappedEvents];
+};
+
+const getEventConversionType = (message) => {
+  const channel = get(message, 'channel');
+  let eventConversionType = message?.properties?.eventConversionType;
+  if (
+    channelMapping[eventConversionType?.toLowerCase()] ||
+    channelMapping[channel?.toLowerCase()]
+  ) {
+    eventConversionType = eventConversionType
+      ? channelMapping[eventConversionType?.toLowerCase()]
+      : channelMapping[channel?.toLowerCase()];
+  } else {
+    eventConversionType = 'OFFLINE';
+  }
+  return eventConversionType;
+};
+
+const validateEventConfiguration = (eventConversionType, pixelId, snapAppId, appId) => {
+  if ((eventConversionType === 'WEB' || eventConversionType === 'OFFLINE') && !pixelId) {
+    throw new ConfigurationError('Pixel Id is required for web and offline events');
+  }
+
+  if (eventConversionType === 'MOBILE_APP' && !(appId && snapAppId)) {
+    let requiredId = 'App Id';
+    if (!snapAppId) {
+      requiredId = 'Snap App Id';
+    }
+    throw new ConfigurationError(`${requiredId} is required for app events`);
+  }
+};
+
+const getEventTimestamp = (message, requiredDays = 37) => {
+  const eventTime = getFieldValueFromMessage(message, 'timestamp');
+  if (eventTime) {
+    const start = moment.unix(moment(eventTime).format('X'));
+    const current = moment.unix(moment().format('X'));
+    // calculates past event in days
+    const deltaDay = Math.ceil(moment.duration(current.diff(start)).asDays());
+    if (deltaDay > requiredDays) {
+      throw new InstrumentationError(
+        `Events must be sent within ${requiredDays} days of their occurrence`,
+      );
+    }
+    return msUnixTimestamp(eventTime)?.toString()?.slice(0, 10);
+  }
+  return eventTime;
+};
+
 module.exports = {
   msUnixTimestamp,
   getItemIds,
@@ -134,6 +214,10 @@ module.exports = {
   getDataUseValue,
   getNormalizedPhoneNumber,
   getHashedValue,
-  channelMapping,
   generateBatchedPayloadForArray,
+  eventMappingHandler,
+  getEventConversionType,
+  validateEventConfiguration,
+  channelMapping,
+  getEventTimestamp,
 };